+++ src/core/wallets/solana_wallet/index.ts
@@ -1,1 +1,2 @@
-export { EdwinSolanaWallet } from './solana_wallet';
+export * from './solana_wallet';
+export * from './solana_public_key_wallet';
