+++ src/plugins/jupiter/jupiterPlugin.ts
@@ -1,11 +1,11 @@
 import { EdwinPlugin } from '../../core/classes/edwinPlugin';
 import { EdwinTool, Chain } from '../../core/types';
 import { JupiterService } from './jupiterService';
-import { EdwinSolanaWallet } from '../../core/wallets';
+import { EdwinSolanaPublicKeyWallet } from '../../core/wallets/solana_wallet';
 import { SwapParametersSchema, SwapParameters, GetTokenAddressSchema, GetTokenAddressParameters } from './parameters';
 
 export class JupiterPlugin extends EdwinPlugin {
-    constructor(wallet: EdwinSolanaWallet) {
+    constructor(wallet: EdwinSolanaPublicKeyWallet) {
         super('jupiter', [new JupiterService(wallet)]);
     }
 
@@ -53,4 +53,4 @@ export class JupiterPlugin extends EdwinPlugin {
     supportsChain = (chain: Chain) => chain.type === 'solana';
 }
 
-export const jupiter = (wallet: EdwinSolanaWallet) => new JupiterPlugin(wallet);
+export const jupiter = (wallet: EdwinSolanaPublicKeyWallet) => new JupiterPlugin(wallet);
