+++ src/plugins/lulo/luloPlugin.ts
@@ -1,11 +1,11 @@
 import { EdwinPlugin } from '../../core/classes/edwinPlugin';
 import { EdwinTool, Chain } from '../../core/types';
 import { LuloProtocol } from './luloProtocol';
-import { EdwinSolanaWallet } from '../../core/wallets';
+import { EdwinSolanaPublicKeyWallet } from '../../core/wallets/solana_wallet';
 import { SupplyParametersSchema, WithdrawParametersSchema, SupplyParameters, WithdrawParameters } from './parameters';
 
 export class LuloPlugin extends EdwinPlugin {
-    constructor(wallet: EdwinSolanaWallet) {
+    constructor(wallet: EdwinSolanaPublicKeyWallet) {
         super('lulo', [new LuloProtocol(wallet)]);
     }
 
@@ -35,4 +35,4 @@ export class LuloPlugin extends EdwinPlugin {
     supportsChain = (chain: Chain) => chain.type === 'solana';
 }
 
-export const lulo = (wallet: EdwinSolanaWallet) => new LuloPlugin(wallet);
+export const lulo = (wallet: EdwinSolanaPublicKeyWallet) => new LuloPlugin(wallet);
